Move geometric GridTools functions to a new file.
grid_tools.cc is presently our most expensive file to compile. This is the first
commit in a sequence which will chop it up in the same way we chopped up
VectorTools.
